Moved types to react-widget, changes to CSS for button state, fixed imports
