+++ packages/react-widget/src/assets/voucherify.css
@@ -121,6 +121,26 @@ div.voucherifyContainer input.voucherifyPublishStatus {
 	text-align: center;
 	outline: none;
 }
+
+div.voucherifyContainer button.voucherifyRedeem,
+div.voucherifyContainer button.voucherifyValidate,
+div.voucherifyContainer button.voucherifyPublish {
+	display: block;
+	margin: 1.5em auto 1.5em auto;
+	border: none;
+	border-radius: 2em;
+	padding: 0 1.75em;
+	cursor: pointer;
+	outline: none;
+	line-height: 3em;
+	height: 3em;
+	font-size: 0.8em;
+	font-weight: bold;
+	text-transform: uppercase;
+	background-color: #08ffb9;
+	color: rgba(63, 49, 49, 0.8);
+}
+
 div.voucherifyContainer input.voucherifyCode.voucherifyInvalid,
 div.voucherifyContainer input.voucherifyAmount.voucherifyInvalid,
 div.voucherifyContainer input.voucherifyCustomerName.voucherifyInvalid,
@@ -131,7 +151,8 @@ div.voucherifyContainer input.voucherifyCustomerLine2.voucherifyInvalid,
 div.voucherifyContainer input.voucherifyCustomerPostalCode.voucherifyInvalid,
 div.voucherifyContainer input.voucherifyCustomerCity.voucherifyInvalid,
 div.voucherifyContainer input.voucherifyCustomerState.voucherifyInvalid,
-div.voucherifyContainer input.voucherifyCustomerCountry.voucherifyInvalid {
+div.voucherifyContainer input.voucherifyCustomerCountry.voucherifyInvalid,
+div.voucherifyContainer button.voucherifyValidate.voucherifyInvalid {
 	background-color: #f8dffe;
 	border: 1px solid #ebd7f4;
 }
@@ -148,7 +169,8 @@ div.voucherifyContainer input.voucherifyCustomerCity.voucherifyValid,
 div.voucherifyContainer input.voucherifyCustomerState.voucherifyValid,
 div.voucherifyContainer input.voucherifyCustomerCountry.voucherifyValid,
 div.voucherifyContainer input.voucherifyPublishStatus.voucherifyValid,
-div.voucherifyContainer button.voucherifyPublish.voucherifyValid {
+div.voucherifyContainer button.voucherifyPublish.voucherifyValid,
+div.voucherifyContainer button.voucherifyValidate.voucherifyValid {
 	border: 1px solid #08ffb9;
 	background-color: #08ffb9;
 	color: rgba(63, 49, 49, 0.8);
@@ -171,36 +193,8 @@ div.voucherifyContainer .voucherifyRow > input:only-child {
 	width: 100%;
 }
 
-div.voucherifyContainer .voucherifyRow > input:not(:last-of-type) {
-	/* margin-right: 0.8em; */
-}
-
-div.voucherifyContainer button.voucherifyRedeem,
-div.voucherifyContainer button.voucherifyValidate,
-div.voucherifyContainer button.voucherifyPublish {
-	display: block;
-
-	margin: 1.5em auto 1.5em auto;
-	border: none;
-	border-radius: 2em;
-	padding: 0 1.75em;
-
-	cursor: pointer;
-	outline: none;
-
-	line-height: 3em;
-	height: 3em;
-
-	font-size: 0.8em;
-	font-weight: bold;
-
-	text-transform: uppercase;
-	background-color: #08ffb9;
-	color: rgba(63, 49, 49, 0.8);
-}
-
-div.voucherifyContainer button.voucherifyValidate:hover,
-div.voucherifyContainer button.voucherifyPublish:hover {
+div.voucherifyContainer button.voucherifyValidate:not(.voucherifyInvalid):hover,
+div.voucherifyContainer button.voucherifyPublish:not(.voucherifyInvalid):hover {
 	background-color: rgba(8, 255, 185, 0.8);
 }
