+++ packages/react-widget/src/types/VoucherifyRedeem.ts
@@ -1,0 +1,12 @@
+export interface VoucherifyRedeemInputsState {
+	[index: string]: boolean
+	voucherifyCode: boolean
+	voucherifyAmount: boolean
+	voucherifyRedeem: boolean
+}
+export interface VoucherifyRedeemInputs {
+	[index: string]: string
+	voucherifyCode: string
+	voucherifyAmount: string
+	voucherifyTracking: string
+}
