+++ packages/sdk/src/types/ClientSide.ts
@@ -99,71 +99,6 @@ export interface ClientSidePublishPayload {
 	metadata?: Record<string, any>
 }
 
-export interface VoucherifyPublishInputs {
-	[index: string]: string
-	name: string
-	email: string
-	phone: string
-	line_1: string
-	line_2: string
-	postal_code: string
-	city: string
-	state: string
-	country: string
-	voucherifyPublishStatus: string
-	voucherifyTracking: string
-	voucherifyPublish: string
-}
-
-export interface VoucherifyPublishInputsState {
-	[index: string]: boolean
-	name: boolean
-	phone: boolean
-	email: boolean
-	line_1: boolean
-	line_2: boolean
-	postal_code: boolean
-	city: boolean
-	state: boolean
-	country: boolean
-	voucherifyPublishStatus: boolean
-	voucherifyPublish: boolean
-}
-
-export interface VoucherifyRedeemInputsState {
-	[index: string]: boolean
-	voucherifyCode: boolean
-	voucherifyAmount: boolean
-	voucherifyRedeem: boolean
-}
-export interface VoucherifyRedeemInputs {
-	[index: string]: string
-	voucherifyCode: string
-	voucherifyAmount: string
-	voucherifyTracking: string
-}
-
-export interface VoucherifyValidateInputsState {
-	[index: string]: boolean
-	voucherifyCode: boolean
-	voucherifyAmount: boolean
-	voucherifyValidate: boolean
-}
-export interface VoucherifyValidateInputs {
-	[index: string]: string | number
-	voucherifyCode: string
-	voucherifyAmount: string
-	voucherifyDiscountType: string
-	voucherifyPercentOff: number
-	voucherifyAmountOff: number
-	voucherifyUnitOff: number
-	voucherifyTracking: string
-}
-export type NotDefinedPlaceholder = Pick<
-	VoucherifyPublishInputs,
-	'name' | 'email' | 'phone' | 'line_1' | 'line_2' | 'postal_code' | 'city' | 'state' | 'country'
-> & { [index: string]: string }
-
 export type ClientSidePublishPreparedPayload = ClientSidePublishPayload
 
 export interface ClientSidePublishQueryParams {
